requirements.txt
aliyun-python-sdk-core >= 2.0.19
aliyun-python-sdk-ecs >= 0.1.4
#aliyun-python-sdk-oss >= 1.0.0
#aliyuncli >= 2.0.6
amqp >= 1.4.7
anyjson >= 0.3.3
billiard >= 3.3.0.21
colorama >= 0.3.3
defusedxml >= 0.4.1
Django >= 1.8.6
django-braces >= 1.8.1
django-celery == 3.2.2
django-cors-headers==2.2.0
django-debug-toolbar >= 1.4
django-filter >= 0.11.0
django-guardian >= 1.3.2
django-haystack >= 2.4.1
django-model-utils >= 2.3.1
django-notifications >= 0.1.dev0
django-notifications-hq >= 0.7.1
django-oauth-toolkit >= 0.9.0
djangorestframework >= 3.3.1
djangorestframework-jsonp >= 1.0.2
djangorestframework-jwt==1.11
djangorestframework-xml >= 1.2.0
djangorestframework-yaml >= 1.0.2
celery-with-redis >= 3.0
dnspython >= 1.12.0
drf-haystack >= 1.5.5
ecdsa >= 0.13
elasticsearch >= 2.1.0
Fabric >= 1.10.2
httplib2 >= 0.9.2
jmespath >= 0.9.0
jsonfield >= 1.0.3
kombu >= 3.0.29
Markdown >= 2.6.4
pymysql>= 0.7.11
oauthlib >= 1.0.3
paramiko >= 1.16.0
passlib >= 1.6.5
pip >= 7.1.2
pycrypto >= 2.6.1
Pygments >= 2.0.2
python-dateutil >= 2.4.2
python-ldap==3.0.0
python-gitlab==1.3.0
pytz >= 2015.7
PyYAML >= 3.11
pyzabbix >= 0.7.3
requests >= 2.8.1
rfc3986 >= 0.3.0
setuptools >= 18.5
six >= 1.10.0
sqlparse >= 0.1.18
urllib3 >= 1.12
wheel >= 0.26.0
ipython==6.2.1
mysqlclient==1.3.4
ldap3==2.4
redis==2.10.6
SQLAlchemy==1.2.7
Jinja2==2.10
drf-writable-nested==0.4.2
django-sequences == 2.2
wrapt == 1.10.11
django-mptt == 0.9.0
pandas == 0.22.0
openpyxl == 2.5.3
xlrd == 1.1.0